requirements.txt
Django~=4.2.4
pandas~=2.0.3
opencv-python~=4.8.0.76
djangorestframework~=3.14.0
django-extensions
graphviz
pyparsing
pydot
django-rest-swagger
drf-yasg